Rename DapDurableObject and GarbageCollector
